+++ examples/cypress/CY8CKIT-064S0S2-4343W/src/memfault_platform_config.h
@@ -1,0 +1,19 @@
+//! @file
+//!
+//! Copyright (c) Memfault, Inc.
+//! See License.txt for details
+//!
+//! Platform overrides for the default configuration settings in the memfault-firmware-sdk.
+//! Default configuration settings can be found in "memfault/config.h"
+
+#pragma once
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+#define MEMFAULT_USE_GNU_BUILD_ID 1
+
+#ifdef __cplusplus
+}
+#endif