+++ examples/cypress/CY8CKIT-064S0S2-4343W/src/memfault_platform_log_config.h
@@ -1,0 +1,19 @@
+//! @file
+//!
+//! Copyright (c) Memfault, Inc.
+//! See License.txt for details
+//!
+// Logging depends on how your configuration does logging. See
+// https://mflt.io/logging-dependency
+
+#pragma once
+
+#include "memfault/core/platform/debug_log.h"
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+#ifdef __cplusplus
+}
+#endif